+++ SurgSim/Graphics/Mesh.h
@@ -78,7 +78,7 @@ public:
 	void initialize(const std::vector<SurgSim::Math::Vector3d>& vertices,
 					const std::vector<SurgSim::Math::Vector4d>& colors,
 					const std::vector<SurgSim::Math::Vector2d>& textures,
-					const std::vector<unsigned int>& triangles);
+					const std::vector<size_t>& triangles);
 };
 
 }; // Graphics
