+++ SurgSim/Physics/Fem2DElementTriangle.cpp
@@ -39,7 +39,7 @@ namespace SurgSim
 namespace Physics
 {
 
-Fem2DElementTriangle::Fem2DElementTriangle(std::array<unsigned int, 3> nodeIds)
+Fem2DElementTriangle::Fem2DElementTriangle(std::array<size_t, 3> nodeIds)
 	: m_restArea(0.0),
 	  m_thickness(0.0)
 {
