+++ SurgSim/Physics/UnitTests/Fem3DRepresentationBilateral3DTests.cpp
@@ -44,15 +44,15 @@ namespace SurgSim
 namespace Physics
 {
 
-static std::shared_ptr<Fem3DElementTetrahedron> makeTetrahedron(unsigned int node0,
-																unsigned int node1,
-																unsigned int node2,
-																unsigned int node3,
+static std::shared_ptr<Fem3DElementTetrahedron> makeTetrahedron(size_t node0,
+																size_t node1,
+																size_t node2,
+																size_t node3,
 																double massDensity,
 																double poissonRatio,
 																double youngModulus)
 {
-	std::array<unsigned int, 4> nodes = {node0, node1, node2, node3};
+	std::array<size_t, 4> nodes = {node0, node1, node2, node3};
 	auto element = std::make_shared<Fem3DElementTetrahedron>(nodes);
 	element->setMassDensity(massDensity);
 	element->setPoissonRatio(poissonRatio);
