+++ SurgSim/Physics/UnitTests/FemRepresentationParametersTest.cpp
@@ -148,7 +148,7 @@ TEST(FemRepresentationParametersTest, BoundaryConditionsTest)
 	EXPECT_EQ(0u, femRepresentationParam->getBoundaryConditions().size());
 
 	// Add all
-	std::vector<unsigned int> data;
+	std::vector<size_t> data;
 	data.push_back(0);
 	data.push_back(1);
 	data.push_back(2);
